+++ mtg_proxy_printer/ui/main_window.py
@@ -31,7 +31,7 @@ from mtg_proxy_printer.ui.current_page_view import CurrentPageView
 from mtg_proxy_printer.ui.document_view import DocumentView
 from mtg_proxy_printer.ui.add_card import AddCardWidget
 from mtg_proxy_printer.ui.dialogs import SavePDFDialog, SaveDocumentAsDialog, LoadDocumentDialog, \
-    AboutMTGProxyPrinterDialog, PrintPreviewDialog, PrintDialog
+    AboutMTGProxyPrinterDialog, PrintPreviewDialog, PrintDialog, DocumentSettingsDialog
 from mtg_proxy_printer.ui.cache_cleanup_wizard import CacheCleanupWizard
 from mtg_proxy_printer.ui.deck_import_wizard import DeckImportWizard
 
@@ -85,6 +85,7 @@ class MainWindow(*inherits_from_ui_file_with_name(f"{layout}_search_layout/main_
         self.settings_changed.connect(self.add_card_widget.update_selected_language)
         self.settings_changed.connect(document.apply_settings)
         self.settings_changed.connect(self.page_view.settings_changed)
+        document.page_layout_changed.connect(self.page_view.settings_changed)
         self.settings_changed.connect(self.offer_re_downloading_card_database)
         self.action_show_toolbar: QAction
         self.action_show_toolbar.setChecked(mtg_proxy_printer.settings.settings["gui"].getboolean("show-toolbar"))
@@ -150,6 +151,7 @@ class MainWindow(*inherits_from_ui_file_with_name(f"{layout}_search_layout/main_
             self.action_new_document,
             self.action_save_as,
             self.action_save_document,
+            self.action_edit_document_settings,
             self.action_compact_document,
             self.action_load_document,
             self.action_print,
@@ -381,8 +383,14 @@ class MainWindow(*inherits_from_ui_file_with_name(f"{layout}_search_layout/main_
             logger.debug("Saved.")
 
     @pyqtSlot()
+    def on_action_edit_document_settings_triggered(self):
+        logger.info("User wants to edit the document settings. Showing the editor dialog")
+        dialog = DocumentSettingsDialog(self.document, self)
+        dialog.exec_()
+
+    @pyqtSlot()
     def on_action_save_as_triggered(self):
-        dialog = SaveDocumentAsDialog(self, self.document)
+        dialog = SaveDocumentAsDialog(self, self.document)  # TODO: Unify argument order. Put the parent at the end
         dialog.exec_()
 
     @pyqtSlot()